Move interceptor to separate folder
