+++ apollo-runtime/src/main/java/com/apollographql/apollo/internal/interceptor/ApolloParseInterceptor.java
@@ -5,8 +5,8 @@ import com.apollographql.android.api.graphql.Response;
 import com.apollographql.android.api.graphql.ResponseFieldMapper;
 import com.apollographql.android.api.graphql.ScalarType;
 import com.apollographql.apollo.internal.cache.normalized.ResponseNormalizer;
-import com.apollographql.apollo.ApolloInterceptor;
-import com.apollographql.apollo.ApolloInterceptorChain;
+import com.apollographql.apollo.interceptor.ApolloInterceptor;
+import com.apollographql.apollo.interceptor.ApolloInterceptorChain;
 import com.apollographql.apollo.CustomTypeAdapter;
 import com.apollographql.apollo.HttpException;
 import com.apollographql.apollo.internal.cache.http.HttpCache;
